1) Added sessions to login. 2) Sessions include roles and will be verfied at pages. 3) Initalized controllers for PCC and PCM. 4) Added functionality for admins to edit and delete users

+++ src/main/java/com/sam2021/database/UserRepository.java
@@ -7,6 +7,7 @@ import java.util.List;
 
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, String> {
-    List<UserEntity> findAllById(Integer id );
+    List<UserEntity> findAllById(Long id );
     List<UserEntity> findByEmail(String email);
+
 }
